+++ examples/shared_state_bot/src/main.rs
@@ -4,6 +4,7 @@ use std::sync::atomic::{AtomicU64, Ordering};
 
 use lazy_static::lazy_static;
 use teloxide::prelude::*;
+use tokio_stream::wrappers::UnboundedReceiverStream;
 
 lazy_static! {
     static ref MESSAGES_TOTAL: AtomicU64 = AtomicU64::new(0);
@@ -18,15 +19,15 @@ async fn run() {
     teloxide::enable_logging!();
     log::info!("Starting shared_state_bot...");
 
-    let bot = Bot::from_env();
+    let bot = Bot::from_env().auto_send();
 
     Dispatcher::new(bot)
-        .messages_handler(|rx: DispatcherHandlerRx<Message>| {
-            rx.for_each_concurrent(None, |message| async move {
+        .messages_handler(|rx: DispatcherHandlerRx<AutoSend<Bot>, Message>| {
+            UnboundedReceiverStream::new(rx).for_each_concurrent(None, |message| async move {
                 let previous = MESSAGES_TOTAL.fetch_add(1, Ordering::Relaxed);
 
                 message
-                    .answer_str(format!("I received {} messages in total.", previous))
+                    .answer(format!("I received {} messages in total.", previous))
                     .await
                     .log_on_error()
                     .await;